--- a/client/request.go
+++ b/client/request.go
@@ -218,6 +218,9 @@ func (r *Request) Param(key string) []string {
 func (r *Request) Params() iter.Seq2[string, []string] {
 	return func(yield func(string, []string) bool) {
 		vals := r.params.Len()
+		if vals == 0 {
+			return
+		}
 		prealloc := make([]string, 2*vals)
 		p := pair{
 			k: prealloc[:0:vals],
@@ -454,6 +457,9 @@ func (r *Request) FormData(key string) []string {
 func (r *Request) AllFormData() iter.Seq2[string, []string] {
 	return func(yield func(string, []string) bool) {
 		vals := r.formData.Len()
+		if vals == 0 {
+			return
+		}
 		prealloc := make([]string, 2*vals)
 		p := pair{
 			k: prealloc[:0:vals],