# Description

1. Saves ~8% time on allocations in `Request::Params()`
```bash
benchstat old-params.txt new-params.txt                                                       
goos: darwin
goarch: arm64
pkg: github.com/gofiber/fiber/v3/client
cpu: Apple M3 Max
                │ old-params.txt │           new-params.txt            │
                │     sec/op     │    sec/op     vs base               │
_Request_Params      151.6n ± 0%   139.3n ± 11%  -8.11% (p=0.000 n=20)

                │ old-params.txt │         new-params.txt         │
                │      B/op      │    B/op     vs base            │
_Request_Params       184.0 ± 0%   184.0 ± 0%  ~ (p=1.000 n=20) ¹
¹ all samples are equal

                │ old-params.txt │           new-params.txt           │
                │   allocs/op    │ allocs/op   vs base                │
_Request_Params       6.000 ± 0%   5.000 ± 0%  -16.67% (p=0.000 n=20)
```

2. Saves ~6% time on allocations in `Request::AllFormData()`
```bash
benchstat old-allformdata.txt new-allformdata.txt                                                       
goos: darwin
goarch: arm64
pkg: github.com/gofiber/fiber/v3/client
cpu: Apple M3 Max
                     │ old-allformdata.txt │        new-allformdata.txt         │
                     │       sec/op        │   sec/op     vs base               │
_Request_AllFormData           151.2n ± 1%   142.4n ± 2%  -5.79% (p=0.000 n=20)

                     │ old-allformdata.txt │      new-allformdata.txt       │
                     │        B/op         │    B/op     vs base            │
_Request_AllFormData            184.0 ± 0%   184.0 ± 0%  ~ (p=1.000 n=20) ¹
¹ all samples are equal

                     │ old-allformdata.txt │        new-allformdata.txt         │
                     │      allocs/op      │ allocs/op   vs base                │
_Request_AllFormData            6.000 ± 0%   5.000 ± 0%  -16.67% (p=0.000 n=20)
```

## Changes introduced

List the new features or adjustments introduced in this pull request. Provide details on benchmarks, documentation updates, changelog entries, and if applicable, the migration guide.

- [ ] Benchmarks: Describe any performance benchmarks and improvements related to the changes.
- [ ] Documentation Update: Detail the updates made to the documentation and links to the changed files.
- [ ] Changelog/What's New: Include a summary of the additions for the upcoming release notes.
- [ ] Migration Guide: If necessary, provide a guide or steps for users to migrate their existing code to accommodate these changes.
- [ ] API Alignment with Express: Explain how the changes align with the Express API.
- [ ] API Longevity: Discuss the steps taken to ensure that the new or updated APIs are consistent and not prone to breaking changes.
- [ ] Examples: Provide examples demonstrating the new features or changes in action.

## Type of change

Please delete options that are not relevant.

- [ ] New feature (non-breaking change which adds functionality)
- [ ] Enhancement (improvement to existing features and functionality)
- [ ] Documentation update (changes to documentation)
- [x] Performance improvement (non-breaking change which improves efficiency)
- [ ] Code consistency (non-breaking change which improves code reliability and robustness)

## Checklist

Before you submit your pull request, please make sure you meet these requirements:

- [ ] Followed the inspiration of the Express.js framework for new functionalities, making them similar in usage.
- [ ] Conducted a self-review of the code and provided comments for complex or critical parts.
- [ ] Updated the documentation in the `/docs/` directory for [Fiber's documentation](https://docs.gofiber.io/).
- [ ] Added or updated unit tests to validate the effectiveness of the changes or new features.
- [ ] Ensured that new and existing unit tests pass locally with the changes.
- [ ] Verified that any new dependencies are essential and have been agreed upon by the maintainers/community.
- [ ] Aimed for optimal performance with minimal allocations in the new code.
- [ ] Provided benchmarks for the new code to analyze and improve upon.

## Commit formatting

Please use emojis in commit messages for an easy way to identify the purpose or intention of a commit. Check out the emoji cheatsheet here: [CONTRIBUTING.md](https://github.com/gofiber/fiber/blob/main/.github/CONTRIBUTING.md#pull-requests-or-commits)